requirements.txt
appdirs==1.4.4
APScheduler==3.8.1
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cssselect==1.1.0
debugpy==1.5.1
distlib==0.3.4
fake-useragent==0.1.11
filelock==3.4.0
Flask==2.0.2
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
mypy==0.910
mypy-extensions==0.4.3
parse==1.19.0
pipenv==2021.11.23
platformdirs==2.4.0
pyee==8.2.2
pyppeteer==0.2.6
pyquery==1.4.3
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
tqdm==4.62.3
types-requests==2.26.1
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
w3lib==1.22.0
websockets==9.1
Werkzeug==2.0.2
zipp==3.6.0
